+++ rows/plugins/plugin_json.py
@@ -1,6 +1,6 @@
 # coding: utf-8
 
-# Copyright 2014-2015 Álvaro Justen <https://github.com/turicas/rows/>
+# Copyright 2014-2016 Álvaro Justen <https://github.com/turicas/rows/>
 #
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU General Public License as published by
@@ -21,24 +21,30 @@ import datetime
 import decimal
 import json
 
+import six
+
 from rows import fields
 from rows.plugins.utils import (create_table, export_data,
                                 get_filename_and_fobj, prepare_to_export)
 
 
 def import_from_json(filename_or_fobj, encoding='utf-8', *args, **kwargs):
-    'Import data from a JSON file'
+    '''Import a JSON file or file-like object into a `rows.Table`
 
-    kwargs['encoding'] = encoding
+    If a file-like object is provided it MUST be open in text (non-binary) mode
+    on Python 3 and could be open in both binary or text mode on Python 2.
+    '''
+
     filename, fobj = get_filename_and_fobj(filename_or_fobj)
 
     json_obj = json.load(fobj, encoding=encoding)
-    field_names = json_obj[0].keys()
+    field_names = list(json_obj[0].keys())
     table_rows = [[item[key] for key in field_names] for item in json_obj]
 
-    data = [field_names] + table_rows
-    meta = {'imported_from': 'json', 'filename': filename, }
-    return create_table(data, meta=meta, *args, **kwargs)
+    meta = {'imported_from': 'json',
+            'filename': filename,
+            'encoding': encoding,}
+    return create_table([field_names] + table_rows, meta=meta, *args, **kwargs)
 
 
 def _convert(value, field_type, *args, **kwargs):
@@ -61,18 +67,26 @@ def _convert(value, field_type, *args, **kwargs):
 
 def export_to_json(table, filename_or_fobj=None, encoding='utf-8', indent=None,
                    *args, **kwargs):
+    '''Export a `rows.Table` to a JSON file or file-like object
+
+    If a file-like object is provided it MUST be open in binary mode (like in
+    `open('myfile.json', mode='wb')`).
+    '''
     # TODO: will work only if table.fields is OrderedDict
 
     fields = table.fields
     prepared_table = prepare_to_export(table, *args, **kwargs)
-    field_names = prepared_table.next()
+    field_names = next(prepared_table)
     data = [{field_name: _convert(value, fields[field_name], *args, **kwargs)
              for field_name, value in zip(field_names, row)}
             for row in prepared_table]
 
     result = json.dumps(data, indent=indent)
+    if type(result) is six.text_type:  # Python 3
+        result = result.encode(encoding)
+
     if indent is not None:
         # clean up empty spaces at the end of lines
-        result = '\n'.join(line.rstrip() for line in result.splitlines())
+        result = b'\n'.join(line.rstrip() for line in result.splitlines())
 
-    return export_data(filename_or_fobj, result)
+    return export_data(filename_or_fobj, result, mode='wb')