+++ tests/tests_plugin_xls.py
@@ -1,6 +1,6 @@
 # coding: utf-8
 
-# Copyright 2014-2015 Álvaro Justen <https://github.com/turicas/rows/>
+# Copyright 2014-2016 Álvaro Justen <https://github.com/turicas/rows/>
 #
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU General Public License as published by
@@ -29,7 +29,7 @@ import mock
 import rows
 import rows.fields as fields
 import rows.plugins.xls
-import utils
+import tests.utils as utils
 
 
 def date_to_datetime(value):
@@ -41,6 +41,7 @@ class PluginXlsTestCase(utils.RowsTestMixIn, unittest.TestCase):
     plugin_name = 'xls'
     file_extension = 'xls'
     filename = 'tests/data/all-field-types.xls'
+    assert_meta_encoding = False
 
     def test_imports(self):
         self.assertIs(rows.import_from_xls, rows.plugins.xls.import_from_xls)
@@ -49,14 +50,16 @@ class PluginXlsTestCase(utils.RowsTestMixIn, unittest.TestCase):
     @mock.patch('rows.plugins.xls.create_table')
     def test_import_from_xls_uses_create_table(self, mocked_create_table):
         mocked_create_table.return_value = 42
-        kwargs = {'encoding': 'test', 'some_key': 123, 'other': 456, }
+        kwargs = {'some_key': 123, 'other': 456, }
         result = rows.import_from_xls(self.filename, **kwargs)
         self.assertTrue(mocked_create_table.called)
         self.assertEqual(mocked_create_table.call_count, 1)
         self.assertEqual(result, 42)
 
         call = mocked_create_table.call_args
-        kwargs['meta'] = {'imported_from': 'xls', 'filename': self.filename, }
+        kwargs['meta'] = {'imported_from': 'xls',
+                          'filename': self.filename,
+                          'sheet_name': 'Sheet1', }
         self.assertEqual(call[1], kwargs)
 
     @mock.patch('rows.plugins.xls.create_table')
@@ -66,13 +69,19 @@ class PluginXlsTestCase(utils.RowsTestMixIn, unittest.TestCase):
         # import using filename
         table_1 = rows.import_from_xls(self.filename)
         call_args = mocked_create_table.call_args_list[0]
-        self.assert_create_table_data(call_args)
+        self.assert_create_table_data(call_args,
+                expected_meta={'imported_from': 'xls',
+                               'filename': self.filename,
+                               'sheet_name': 'Sheet1',})
 
         # import using fobj
         with open(self.filename, 'rb') as fobj:
             table_2 = rows.import_from_xls(fobj)
             call_args = mocked_create_table.call_args_list[1]
-            self.assert_create_table_data(call_args)
+            self.assert_create_table_data(call_args,
+                expected_meta={'imported_from': 'xls',
+                               'filename': self.filename,
+                               'sheet_name': 'Sheet1',})
 
     def test_export_to_xls_filename(self):
         # TODO: may test file contents